% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powertest.R
\name{getonesamplesizebypwr}
\alias{getonesamplesizebypwr}
\title{pwr方法暂不支持有非劣或等效界值计算
该函数需要pwr包}
\usage{
getonesamplesizebypwr(alpha = 0.05, power = 0.8, pt, p, sigma = NULL, type)
}
\arguments{
\item{alpha}{第一类错误，默认0.05，默认双侧检验}

\item{power}{检验效能，默认0.8}

\item{pt}{pt实验组概率或均值}

\item{p}{目标值}

\item{sigma}{标准差}

\item{type}{实验类型，只能为Quantitative或Qualitative}
}
\value{
每一类的样本量n
}
\description{
pwr方法暂不支持有非劣或等效界值计算
该函数需要pwr包
}
